+++ docs/CHANGELOG.md
@@ -32,6 +32,7 @@ VoxConvert:
 
    - Fixed `--input` handling if the value is a directory
    - Added option `--wildcard` to specify a wildcard in cases where the `--input` value is a directory
+   - Added `--surface-only` parameter that removes any non surface voxel.
 
 VoxEdit: